+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+v3.0.74
+----------
+ * reduce number of lookup to DB
+
+v3.0.73
+----------
+ * Add test case for search URL against empty field value
+ * Fix sending vumi messages initiated from RapidPro without response to
+
 v3.0.72
 ----------
  * Improvements to external channels to allow configuration against JSON and XML endpoints